+++ packages/devtools_extensions/example/foo/packages/foo/lib/foo.dart
@@ -3,3 +3,4 @@
 // found in the LICENSE file.
 
 export 'src/foo_base.dart';
+export 'src/foo_controller.dart';
