+++ packages/devtools_extensions/example/foo/packages/foo_devtools_extension/lib/src/foo_devtools_extension.dart
@@ -3,10 +3,11 @@
 // found in the LICENSE file.
 
 import 'package:devtools_app_shared/ui.dart';
-import 'package:devtools_extensions/api.dart';
 import 'package:devtools_extensions/devtools_extensions.dart';
 import 'package:flutter/material.dart';
 
+import 'devtools_event_example.dart';
+import 'devtools_extension_api_example.dart';
 import 'service_extension_example.dart';
 
 class FooDevToolsExtension extends StatelessWidget {
@@ -45,117 +46,10 @@ class _FooExtensionHomePageState extends State<FooExtensionHomePage> {
             SizedBox(height: 32.0),
             ListeningForDevToolsEventExample(),
             SizedBox(height: 32.0),
-            CallingDevToolsAPIsExample(),
+            CallingDevToolsExtensionsAPIsExample(),
           ],
         ),
       ),
-    );
-  }
-}
-
-class ListeningForDevToolsEventExample extends StatefulWidget {
-  const ListeningForDevToolsEventExample({super.key});
-
-  @override
-  State<ListeningForDevToolsEventExample> createState() =>
-      _ListeningForDevToolsEventExampleState();
-}
-
-class _ListeningForDevToolsEventExampleState
-    extends State<ListeningForDevToolsEventExample> {
-  String? message;
-
-  @override
-  void initState() {
-    super.initState();
-    // Example of the devtools extension registering a custom handler to listen
-    // for an event coming from DevTools.
-    extensionManager.registerEventHandler(
-      DevToolsExtensionEventType.unknown,
-      (event) {
-        setState(() {
-          message = event.data?.toString() ?? 'unknown event';
-        });
-      },
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          '2. Example of listening for a DevTools event',
-          style: Theme.of(context).textTheme.titleMedium,
-        ),
-        const PaddedDivider.thin(),
-        Text('Received an unknown event from DevTools: $message'),
-      ],
-    );
-  }
-}
-
-class CallingDevToolsAPIsExample extends StatelessWidget {
-  const CallingDevToolsAPIsExample({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          '3. Example of calling DevTools extension APIs',
-          style: Theme.of(context).textTheme.titleMedium,
-        ),
-        const PaddedDivider.thin(),
-        ElevatedButton(
-          onPressed: () => extensionManager.postMessageToDevTools(
-            DevToolsExtensionEvent(
-              DevToolsExtensionEventType.unknown,
-              data: {'foo': 'bar'},
-            ),
-          ),
-          child: const Text('Send a message to DevTools'),
-        ),
-        const SizedBox(height: 16.0),
-        ElevatedButton(
-          onPressed: () =>
-              extensionManager.showNotification('Yay, DevTools Extensions!'),
-          child: const Text('Show DevTools notification'),
-        ),
-        const SizedBox(height: 16.0),
-        ElevatedButton(
-          onPressed: () => extensionManager.showBannerMessage(
-            key: 'example_message_single_dismiss',
-            type: 'warning',
-            message: 'Warning: with great power, comes great '
-                'responsibility. I\'m not going to tell you twice.\n'
-                '(This message can only be shown once)',
-            extensionName: 'foo',
-          ),
-          child: const Text(
-            'Show DevTools warning (ignore if already dismissed)',
-          ),
-        ),
-        const SizedBox(height: 16.0),
-        ElevatedButton(
-          onPressed: () => extensionManager.showBannerMessage(
-            key: 'example_message_multi_dismiss',
-            type: 'warning',
-            message: 'Warning: with great power, comes great '
-                'responsibility. I\'ll keep reminding you if you '
-                'forget.\n(This message can be shown multiple times)',
-            extensionName: 'foo',
-            ignoreIfAlreadyDismissed: false,
-          ),
-          child: const Text(
-            'Show DevTools warning (can show again after dismiss)',
-          ),
-        ),
-      ],
     );
   }
 }
